refactor import statements to improve module structure and clarity
